changed variable type to Long for student_master_id attribute.

+++ smsrestservices/src/main/java/com/sms/model/StudentMaster.java
@@ -17,7 +17,7 @@ public class StudentMaster {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private int student_master_id ;
+	private Long student_master_id ;
 	
 	@Column(name = "first_name")
 	private String fisrtName ;
